+++ packages/api/src/praise/transformers.ts
@@ -88,6 +88,7 @@ const praiseDocumentToDto = async (
     quantifications,
     giver,
     receiver,
+    forwarder,
     createdAt,
     updatedAt,
   } = praiseDocument;
@@ -99,6 +100,7 @@ const praiseDocumentToDto = async (
     quantifications: await quantificationListTransformer(quantifications),
     giver: userAccountTransformer(giver),
     receiver: userAccountTransformer(receiver),
+    forwarder: forwarder ? userAccountTransformer(forwarder) : undefined,
     createdAt: createdAt.toISOString(),
     updatedAt: updatedAt.toISOString(),
   };