+++ packages/api/src/user/types.ts
@@ -5,6 +5,7 @@ export enum UserRole {
   ADMIN = 'ADMIN',
   USER = 'USER',
   QUANTIFIER = 'QUANTIFIER',
+  FORWARDER = 'FORWARDER',
 }
 
 export interface User {